requirements.txt
pytorch-lightning
ipykernel
matplotlib
pandas
torchvision
nilearn
torcheval
plotly
open_clip_torch
transformers
diffusers==0.26.3
wandb
peft
lpips


ipykernel